+++ integration_test/index.test.js
@@ -1,5 +1,8 @@
 import { AppRegistry } from 'react-native';
 import TestableApp from './src/TestableApp';
-import { name as appName } from './app.json';
+import PlayerTestWorld from './playertesting/PlayerTestWorld';
+import { name as appName, licenseKey } from './app.json';
+
+PlayerTestWorld.defaultLicenseKey = licenseKey;
 
 AppRegistry.registerComponent(appName, () => TestableApp);