+++ integration_test/src/App.tsx
@@ -1,121 +1,63 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- */
-
 import React from 'react';
-import type { PropsWithChildren } from 'react';
-import {
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  Text,
-  useColorScheme,
-  View,
-} from 'react-native';
+import { SafeAreaView, StyleSheet, Text } from 'react-native';
 
-import {
-  Colors,
-  DebugInstructions,
-  Header,
-  LearnMoreLinks,
-  ReloadInstructions,
-} from 'react-native/Libraries/NewAppScreen';
-
-type SectionProps = PropsWithChildren<{
-  title: string;
-}>;
-
-function Section({ children, title }: SectionProps): JSX.Element {
-  const isDarkMode = useColorScheme() === 'dark';
-  return (
-    <View style={styles.sectionContainer}>
-      <Text
-        style={[
-          styles.sectionTitle,
-          {
-            color: isDarkMode ? Colors.white : Colors.black,
-          },
-        ]}
-      >
-        {title}
-      </Text>
-      <Text
-        style={[
-          styles.sectionDescription,
-          {
-            color: isDarkMode ? Colors.light : Colors.dark,
-          },
-        ]}
-      >
-        {children}
-      </Text>
-    </View>
-  );
-}
+import { Colors } from 'react-native/Libraries/NewAppScreen';
 
 function App(): JSX.Element {
-  const isDarkMode = useColorScheme() === 'dark';
+  console.info(`
+To execute tests, run one of the following commands:
+- Run on iOS:
+  yarn integration-test test:ios
 
-  const backgroundStyle = {
-    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
-  };
-
+- Run on Android:
+  yarn integration-test test:android
+  `);
   return (
-    <SafeAreaView style={backgroundStyle}>
-      <StatusBar
-        barStyle={isDarkMode ? 'light-content' : 'dark-content'}
-        backgroundColor={backgroundStyle.backgroundColor}
-      />
-      <ScrollView
-        contentInsetAdjustmentBehavior="automatic"
-        style={backgroundStyle}
-      >
-        <Header />
-        <View
-          style={{
-            backgroundColor: isDarkMode ? Colors.black : Colors.white,
-          }}
-        >
-          <Section title="Step One">
-            Edit <Text style={styles.highlight}>App.tsx</Text> to change this
-            screen and then come back to see your edits.
-          </Section>
-          <Section title="See Your Changes">
-            <ReloadInstructions />
-          </Section>
-          <Section title="Debug">
-            <DebugInstructions />
-          </Section>
-          <Section title="Learn More">
-            Read the docs to discover what to do next:
-          </Section>
-          <LearnMoreLinks />
-        </View>
-      </ScrollView>
+    <SafeAreaView style={styles.container}>
+      <Text style={styles.headerText}>
+        To execute tests, run one of the following commands:{'\n'}
+      </Text>
+      <Text style={styles.instructions}>
+        - Run on iOS:{'\n'}
+        <Text style={styles.code}>yarn integration-test test:ios</Text>
+      </Text>
+      <Text style={styles.spacer}>{'\n'}</Text>
+      <Text style={styles.instructions}>
+        - Run on Android:{'\n'}
+        <Text style={styles.code}>yarn integration-test test:android</Text>
+      </Text>
+      <Text style={styles.spacer}>{'\n'}</Text>
+      <Text style={styles.note}>Note: See console for copy-ready commands</Text>
     </SafeAreaView>
   );
 }
 
 const styles = StyleSheet.create({
-  sectionContainer: {
-    marginTop: 32,
-    paddingHorizontal: 24,
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'stretch',
+    margin: 16,
   },
-  sectionTitle: {
+  headerText: {
+    fontSize: 30,
+    color: Colors.darker,
+  },
+  instructions: {
     fontSize: 24,
-    fontWeight: '600',
+    color: Colors.darker,
   },
-  sectionDescription: {
-    marginTop: 8,
-    fontSize: 18,
-    fontWeight: '400',
+  code: {
+    fontSize: 20,
+    color: 'gray',
+    fontStyle: 'italic',
   },
-  highlight: {
-    fontWeight: '700',
+  note: {
+    fontSize: 14,
+    color: 'gray',
+  },
+  spacer: {
+    height: 10,
   },
 });